bugfix(DPAV-1151): logout on query 401/403
