+++ frontend/src/utils/Map/loadLocations.ts
@@ -9,10 +9,9 @@ import { Search } from '../Search';
 import { type AsyncReturnType } from '../types';
 
 type CallbackType = (options: AsyncReturnType<typeof Search.location>) => void;
-type Type = (inputValue: string | Coordinates, callback: CallbackType) => void;
 
 function searchLocation(inputValue: string | Coordinates, callback: CallbackType) {
   Search.location(inputValue).then((options) => callback(options));
 }
 
-export const loadLocations: Type = debounce(searchLocation, 500);
+export const loadLocations = debounce(searchLocation, 500);
