+++ frontend/src/utils/debounce.ts
@@ -2,15 +2,13 @@
 // © Crown Copyright 2025. This work has been developed by the National Digital Twin Programme
 // and is legally attributed to the Department for Business and Trade (UK) as the governing entity.
 
-// eslint-disable-next-line @typescript-eslint/ban-types
-export function debounce(fn: Function, delay: number = 250) {
-  // eslint-disable-next-line no-undef
-  let timeout: NodeJS.Timeout;
+export function debounce<T extends (...args: never[]) => void>(fn: T, delay: number = 250): T {
+  let timeout: ReturnType<typeof setTimeout>;
 
-  return (...args: unknown[]) => {
+  return ((...args: Parameters<T>) => {
     clearTimeout(timeout);
     timeout = setTimeout(() => {
       fn(...args);
     }, delay);
-  };
+  }) as T;
 }
